pyproject.toml
[project]
name = "emergenciavehicular-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi[all]>=0.135.3",
    "passlib>=1.7.4",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]
